refactor(config): widen operator id type to 32-bit
Replace uint16_t with uint32_t for operator identifiers and update
findOperator accordingly. Parse scanned input with strtoul instead
of atoi to avoid truncation of large ids, and add a fallback that
reflects the raw id on the numeric fields with a "PAGE?" label when
the current page is not a recognized login page.

+++ firmware/CaesarFirmwareV1/config.h
@@ -10,7 +10,7 @@
 #define SCANNER_BAUD 9600
 
 struct Operator {
-  uint16_t id;
+  uint32_t id;
   const char* name;
 };
 
@@ -22,7 +22,7 @@ static const Operator OPERATORS[] = {
 
 static const size_t OPERATOR_COUNT = sizeof(OPERATORS) / sizeof(OPERATORS[0]);
 
-inline const Operator* findOperator(uint16_t id) {
+inline const Operator* findOperator(uint32_t id) {
   for (size_t i = 0; i < OPERATOR_COUNT; i++) {
     if (OPERATORS[i].id == id) {
       return &OPERATORS[i];
